# Improve implementation of `extractKeys`. Now ~30% faster.