--- a/libs/cardano-data/test/Main.hs
+++ b/libs/cardano-data/test/Main.hs
@@ -1,6 +1,7 @@
 module Main where
 
 import Test.Data.Coders (codersTest)
+import Test.Data.MapExtras (mapExtrasTests)
 import Test.Data.UMap (alltests)
 import Test.Tasty
 
@@ -10,7 +11,8 @@ tests :: TestTree
 tests =
   testGroup
     "cardano-data"
-    [ alltests,
+    [ mapExtrasTests,
+      alltests,
       codersTest
     ]