* Add tests and benchmarks for `extractKeys`
* Slight improvement to `splitMember` and addition of a better version for `Set`

Some benchmark results for the record:

* `extractKeysFoldSet` is the previous implementation
* `extractKeysNaive m s === (withoutKeys m s, restrictKeys m s)`

```
benchmarking DeleteTxOuts/extractKeys
time                 121.3 ms   (120.6 ms .. 121.9 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 122.1 ms   (121.6 ms .. 123.4 ms)
std dev              1.186 ms   (328.2 μs .. 1.897 ms)
variance introduced by outliers: 11% (moderately inflated)

benchmarking DeleteTxOuts/extractKeysFoldSet
time                 172.7 ms   (172.1 ms .. 173.3 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 173.2 ms   (172.8 ms .. 174.5 ms)
std dev              1.079 ms   (101.6 μs .. 1.573 ms)
variance introduced by outliers: 12% (moderately inflated)

benchmarking DeleteTxOuts/extractKeysNaive
time                 224.5 ms   (221.4 ms .. 228.1 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 223.2 ms   (221.6 ms .. 224.8 ms)
std dev              2.305 ms   (1.383 ms .. 3.325 ms)
variance introduced by outliers: 14% (moderately inflated)
```